Update README to 2.5 Rev (#62)
* Update version history

Updates references to 2.5 Rev and changes "standard" to "version"

* Add version 2.5 to Do Not Use

Adding the previous VPAT version 2.5 to the section with versions not to be used anymore for new documents

* Correct version typo

Typo in version 2.4 Rev is now no more!

* Update latest VPAT version

Updates the latest VPAT version details to 2.5 Rev and changes the description of what is new

* Tidy version history

Removing all versions of VPAT that are no longer used in product VPATs. Keeping the latest version in use (2.5 Rev)  and 1.0 which BMC and SpringerOpen still use.

* Tidy version update notes

Removes the update notes for VPAT versions we are no longer using, keeps 1.0 as BMC and SpringerOpen still use it

* Tidy conformity wording

Co-authored-by: hollsk <1948361+hollsk@users.noreply.github.com>

* Change "should" to "must"

Because new documents *must* conform!

Co-authored-by: hollsk <1948361+hollsk@users.noreply.github.com>

---------

Co-authored-by: hollsk <1948361+hollsk@users.noreply.github.com>