Springer Nature Link VPAT language corrections (#64)
* Update product name in summary

Changes the product name from SpringerLink to Springer Nature Link in the summary section only

* Fix broken links in summary

Fixes the links for ordering products and the FE playbook accessibility tools,

* Change from SNL to "this product"

Makes the references to the product consistent, keeping "Springer Nature Link" only in the VPAT summary, and "this product" in the rest of the document. This helps future-proof the VPAT against human error in the case of future rebrands and keeps consistent language.

* Capitalise Non-web document

A tidy change to capitalise "Non-web document", because it was bothering me!

* Update SN order link

The SN order link was broken, so changed to the current link /public/cart

* Fix WCAG report anchor link

Fixes the broken anchor link for the WCAG report section within the VPAT

* Revert "Fix WCAG report anchor link"

This reverts commit 14ffd3e59b844dbcb13617df36a42fef1f692136.

* Fix WCAG report anchor links

Fixes the broken anchor links to the WCAG report section within the VPAT

* Fix Section 508 links

Fixes the broken links to section 508 chapters on the US Access Board site

* Fix homepage link in text spacing section

* Change file name

Changes the file name from springerlink.md to springer-nature-link.md to reflect the new product name

* Fix broken 3.1 Definitions link

* Amend references to checkout pages

Removes the stated link for checkout process pages. Keeps the references general as "checkout process pages", which will help keep things tidy given we have more than 1 checkout flow on the product.

* Update report date

Changes the report date to reflect the latest date changes have been made, and re-adds the line break between Notes and the next line